Update Liberland-constitution.md
The constitutional draft was too negative for my taste and did not promote the positive aspects of what is to be. The understatement of blame and dismay about the existing systems has no place in my mind, when creating a political revolutionary state system. The mix of where the items are placed are also incorrect in my mind. The constitution should be minimal – some listed legal issues belong partially in a process regulation ordinance.
The wording was written as any other legal document, where I think it must reflect that the solution to many problems is to write such a document for all to understand without legalese. Simple and straight to the point.
I did not pursue the constitutional law and governmental organization. I shall review the Bill of rights still as it would double items from here and seems to complicated in wording and reading